# swagger.yaml
swagger: "2.0"
info:
    title: Task Manager API
    description: Manage your tasks.
    version: 1.0.0
host: jeffvh-task-manager.herokuapp.com
# host: localhost:3000
basePath: /
schemes:
    - https
paths:
    /users:
        post:
            summary: Create a new user.
            produces:
                - application/json
            responses:
                201:
                    description: Created
                400:
                    description: Bad request
    /users/login:
        post:
            summary: Login with credentials.
            description: Returns a Bearer token.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                400:
                    description: Bad request
    /users/logout:
        post:
            summary: Logout of application from current browser.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                500:
                    description: Server error
    /users/logoutAll:
        post:
            summary: Logout of application from all locations.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                500:
                    description: Server error
    /users/me:
        get:
            summary: Returns informaiton for logged in user.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
        patch:
            summary: Update logged in user's details.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                500:
                    description: Server error
        delete:
            summary: Update logged in user's details.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                500:
                    description: Server error
    /users/me/avatar:
        post:
            summary: Upload an avatar for own account.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                400:
                    description: Bad request
        delete:
            summary: Remove avatar for authenticated account.
            description: Need to be authenticated.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                500:
                    description: Server error
    /users/{id}/avatar:
        get:
            summary: Get avatar for user with specified ID.
            parameters:
                - in: path
                  name: id
                  required: true
                  type: integer
                  minimum: 1
                  description: The id of the user.
            produces:
                - application/json
            responses:
                200:
                    description: OK
                404:
                    description: User with specified ID was not found
    /tasks:
        post:
            summary: Create a new task.
            description: authentication required
            produces:
                - application/json
            responses:
                201:
                    description: Created
                400:
                    description: Bad request
        get:
            summary: Get a list of tasks.
            description: authentication required
            parameters:
                - in: query
                  name: completed
                  required: false
                  type: boolean
                  description: If specified, will search for completed or non-completed tasks. Leave blank to search for all.
                - in: query
                  name: limit
                  required: false
                  type: integer
                  description: max number of tasks to get.
                - in: query
                  name: skip
                  required: false
                  type: integer
                  description: number of tasks to skip.
                - in: query
                  name: sortBy
                  required: false
                  type: string
                  description: in format <property>:<sort method>. Example => sortBy=createdAt:desc
            produces:
                - application/json
            responses:
                200:
                    description: Ok
                500:
                    description: Error gettings tasks
    /tasks/{id}:
        get:
            summary: Get a task.
            description: authentication required
            parameters:
                - in: path
                  name: id
                  required: true
                  type: integer
                  description: id of the task to obtain.
            produces:
                - application/json
            responses:
                200:
                    description: Ok
                500:
                    description: Error getting task
        patch:
            summary: update a task
            description: authentication required
            parameters:
                - in: path
                  name: id
                  required: true
                  type: integer
                  description: id of the task to alter.
            produces:
                - application/json
            responses:
                200:
                    description: Ok
                404:
                    description: Task not found
                500:
                    description: Error with updating task
        delete:
            summary: delete a task
            description: authentication required
            parameters:
                - in: path
                  name: id
                  required: true
                  type: integer
                  description: id of the task to delete.
            produces:
                - application/json
            responses:
                200:
                    description: Ok
                404:
                    description: Task not found
                500:
                    description: Error with updating task